requirements.txt
pandas>=2.0
openpyxl>=3.1
flask>=3.0
pyyaml>=6.0
tqdm>=4.66

playwright>=1.43
requests>=2.31
